pyproject.toml
[project]
name = "nancy-bot"
version = "0.1.0"
description = "Nancy - AI assistant for microlensing research"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.8.0",
    "slack-sdk>=3.18.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=5.3",
    "txtai>=7.0.0",
    "requests>=2.28.0",
    "torch>=1.12.1",
    "faiss-cpu>=1.7.1.post2",
    "transformers>=4.45.0",
    "regex>=2022.8.17"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
